move searching tests out into separate file
